Create a search form component
- Extract the search form view and logic from the top header and create it as a separate component.
