+++ src/components/index.js
@@ -1,5 +1,6 @@
 import "./common/index.js";
 import "./TopHeader.js";
+import "./SearchForm.js";
 import "./AutocompletePanel.js";
 import "./InfiniteCarousel.js";
 import "./CardsPanel.js";
