+++ src/styles/components/TopHeader.scss
@@ -41,36 +41,6 @@ header {
 
   #search-container {
     flex-grow: 1;
-
-    form {
-      width: 100%;
-      display: flex;
-      position: relative;
-
-      input {
-        height: 40px;
-        flex-grow: 1;
-        padding-left: 12px;
-        border-top-left-radius: 4px;
-        border-bottom-left-radius: 4px;
-
-        &:focus ~ autocomplete-panel {
-          opacity: 1;
-          visibility: visible;
-        }
-      }
-
-      button {
-        width: 40px;
-        height: 40px;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        background-color: $color-primary-orange-100;
-        border-top-right-radius: 4px;
-        border-bottom-right-radius: 4px;
-      }
-    }
   }
 
   #lang-link {
